Initial version of CONTRIBUTING.md
I wrote this initial version of CONTRIBUTING.md file. Please review it and feel free to edit.

According to the Milestone 1 assignment description:

"- How others, outside the core team, can contribute to your project somewhere in your repository. This is typically done as a separate file in the repository called `CONTRIBUTING.md`. 

- Here are some examples of this file:
    - [Altair](https://github.com/altair-viz/altair/blob/master/CONTRIBUTING.md)
    - [dplyr](https://github.com/tidyverse/dplyr/blob/master/.github/CONTRIBUTING.md)
    - [Atom editor](https://github.com/atom/atom/blob/master/CONTRIBUTING.md) (very comprehensive)

Note 1 - this document should live in the root of your public GitHub.com repository.

Note 2 - If you all, or even part, of someone else’s code of conduct, be sure to make this clear by attributing them."