report.ipynb USD convertion
Feedback addressed: peer review. "Uncertainty as to the amount of data cleaning conducted for the project. In your report, you state that the original dataset has all of the prices listed as Indian Rupees. However, later on in the report, the prices are presented as USD. There was no mention of whether prices were converted prior to conducting the analysis, nor what the conversion factor that was used (as currency conversions change day-to-day), which does lend some uncertainty as to the validity of the results. From looking at your cleaned training dataset, you have a HP Envy x360 listed as being priced at $96,590, which lends confusion as to when everything was converted or if at all."